made the code compile by fixing errors. fixed various syntax and formatting errors
the code now compiles. however there is no output this could be for a variety of reasons. possibly something to do with memory. there could be more than on instance of this happening. TODO has been left in order to guide you. comment out tests and code until you find the issue and then fix it. good luck